+++ lib/debounced/no_server_error.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+module Debounced
+  class NoServerError < StandardError
+  end
+end